+++ Changelog.md
@@ -3,5 +3,16 @@
 ______________________________________
 ## [Unreleased]
 
+## [1.0.1] - 7-7-2021
+
+
+______________________________________
+
+## [Released]
+
 ## [1.0.0] - 7-6-2021
 ### Added
+
+/// selenium node modules with dependencies for later use to expand the bots use.
+
+/// "!autoungate" command now creates an embedded link that brings you straight to the amazon seller page and that item so you can easily hit "sell as new" and then "apply to sell" for a quick and painless auto ungate. this commands format is "!autoungate (asin)" if you try to send no asin it will tell you how to aquire an asin.